+++ changelog.md
@@ -16,6 +16,7 @@ This is what we are currently working on:
   - Fixed missing hitlocation modifiers
   - Add Quick Note under Encumbrance
   - Added support for Parry / Block modifiers
+  - 'Grand' unification of attribute/skill parsing
 
 0.8.17 - 3/8/2021
   - Bug fix for names/names containing '%'